pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[project]
name = "antistasi_translation_sync"
authors = [{ name = "antistasi_tools" }]
classifiers = ["License :: OSI Approved :: MIT License"]
license = { file = "LICENSE" }
readme = { file = "README.rst", content-type = "text/x-rst" }
dynamic = ['version', 'description']
requires-python = ">=3.11"
dependencies = ["python-dotenv", "httpx~=0.24"]

[project.urls]
Source = "https://github.com/Giddius/Antistasi_Translation_Sync"

[project.scripts]
antistasi-translation-sync = "antistasi_translation_sync.__main__:main"

[project.optional-dependencies]


[tool.flit.sdist]
exclude = [
    "misc",
    "temp",
    "docs",
    "tools",
    ".pydeps",
    "tests",
    ".gitattributes",
    ".gitignore",
    ".editorconfig",
    "planning.todo",
    "tasks.py",
]
[tool.setuptools]
packages = ["gidapptools"]

[tool.autopep8]
ignore = ["E402", "E501", "E731"]
max_line_length = 250

[tool.isort]
balanced_wrapping = true
ensure_newline_before_comments = true
group_by_package = true
import_heading_firstparty = "* Local Imports --------------------------------------------------------------------------------------->"
import_heading_gid = "* Gid Imports ----------------------------------------------------------------------------------------->"
import_heading_qt = "* Qt Imports --------------------------------------------------------------------------------------->"
import_heading_stdlib = "* Standard Library Imports ---------------------------------------------------------------------------->"
import_heading_thirdparty = "* Third Party Imports --------------------------------------------------------------------------------->"
known_gid = ["gid*"]
known_qt = ["PyQt5", "PyQt6", "PySide6", "pyqtgraph"]
length_sort = true
line_length = 200
lines_after_imports = 1
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "QT",
    "GID",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
indented_import_headings = false

[tool.autoflake]
# additional_imports=None
expand_star_imports = true
remove_all_unused_imports = true
remove_duplicate_keys = false
remove_unused_variables = false
ignore_init_module_imports = false


[tool.pytest.ini_options]
addopts = "-rA -vv --durations=10 --capture=tee-sys --show-capture=all --color=auto --code-highlight=yes --no-cov"
norecursedirs = "manual_*"


[tool.mypy]


[tool.ruff]
ignore = ["E501", "E402", "E731"]
